+++ packages/emoji-picker/src/GroupListHeader.tsx
@@ -22,6 +22,7 @@ export interface GroupListHeaderProps {
   group: GroupKey;
   onClear: () => void;
   skinTonePalette?: React.ReactNode;
+  sticky?: boolean;
 }
 
 export class GroupListHeader extends React.PureComponent<GroupListHeaderProps & WithContextProps> {
@@ -32,11 +33,13 @@ export class GroupListHeader extends React.PureComponent<GroupListHeaderProps & 
     group: PropTypes.string.isRequired,
     onClear: PropTypes.func.isRequired,
     skinTonePalette: PropTypes.node,
+    sticky: PropTypes.bool,
   };
 
   static defaultProps = {
     clearIcon: null,
     skinTonePalette: null,
+    sticky: false,
   };
 
   private handleClear = (event: React.MouseEvent<HTMLButtonElement>) => {
@@ -52,6 +55,7 @@ export class GroupListHeader extends React.PureComponent<GroupListHeaderProps & 
       context: { classNames, messages },
       group,
       skinTonePalette,
+      sticky,
     } = this.props;
     const showClear = clearIcon && group === GROUP_KEY_COMMONLY_USED;
     const showPalette =
@@ -59,9 +63,14 @@ export class GroupListHeader extends React.PureComponent<GroupListHeaderProps & 
       (group === GROUP_KEY_SMILEYS_PEOPLE ||
         group === GROUP_KEY_SEARCH_RESULTS ||
         group === GROUP_KEY_NONE);
+    const className = [classNames.emojisHeader];
+
+    if (sticky) {
+      className.push(classNames.emojisHeaderSticky);
+    }
 
     return (
-      <header className={classNames.emojisHeader}>
+      <header className={className.join(' ')}>
         <span>
           {group === GROUP_KEY_COMMONLY_USED
             ? messages[camelCase(commonMode)]