+++ packages/emoji-picker/CHANGELOG.md
@@ -15,7 +15,9 @@
 
 - Added the ability to clear commonly used emoji. Can be customized with the `clearIcon` prop.
 - Added titles/messages for skin tones within the palette.
+- Added a `stickyGroupHeader` prop to enable sticky group headers.
 - Added an `emojiLoading` class name. Can be used to style a loading state for each emoji.
+- Added an `emojisHeaderSticky` class name. Can be used to style stickied group headers.
 
 #### 🛠 Internal